Yummy pull

+++ apps/views.py
@@ -1,8 +1,8 @@
 """main module that runs the application"""
-from datetime import datetime, timedelta
+from datetime import datetime
+from flask import request, jsonify, g, make_response, json
 import jwt
 from apps import app, auth
-from flask import request, abort, jsonify, g, make_response, json
 from apps.user import Users
 from apps.category import Category, CategorySchema
 from apps.recipe import Recipe, RecipeSchema
@@ -17,7 +17,7 @@ recipe_schema = RecipeSchema()
 recipe_schema = RecipeSchema(many=True)
 
 @app.route('/recipe/api/v1.0/user', methods=['POST'])
-def new_user():
+def register_new_user():
     """function to create new user
     ---
     tags:
@@ -33,20 +33,29 @@ def new_user():
       202:
         description: Username already exist!
     """
+
     firstname = str(request.json.get('firstname', "")).strip()
     lastname = str(request.json.get('lastname', "")).strip()
     username = str(request.json.get('username', "")).strip()
     password = str(request.json.get('password', "")).strip()
 
     if firstname and lastname and username and password:
-        user_exit = Users.query.filter_by(username=username).first()
-        if user_exit:
-            response = jsonify({"message":"username {} already exits!".format(username)}, 202)
+        if firstname.isalpha() and lastname.isalpha() and username.isalpha():
+            if password.isalpha():
+                return jsonify({"Message":"Password can't be only alphabets"}), 400
+
+            user_exit = Users.query.filter_by(username=username).first()
+            if user_exit:
+                return jsonify({"message":"username {} already exits!".format(username)}), 400
+            else:
+                user = Users(firstname, lastname, username, password)
+                user.save()
+                response = jsonify({'message': "user {} registered successfully!".format(user.username)}), 201
+                return response
         else:
-            user = Users(firstname, lastname, username, password)
-            user.save()
-            response = jsonify({'message': "user {} registered successfully!".format(user.username)}, 201)
-    return response
+            return jsonify({"Message":"Only password can contain none alphabet character"}), 400
+
+    return jsonify({"Message":"Please fill in all fields"}), 400
 
 
 @app.route('/recipe/api/v1.0/user', methods=['GET'])
@@ -61,29 +70,10 @@ def view_users():
     """
     userlist = Users.getusers()
 
-    response = jsonify(userlist, 200)
-
-    return response
-
-
-@app.route('/recipe/api/v1.0/user/login', methods=['POST'])
-def login_user():
-    """function to verify username and password"""
-    username = request.json.get('username', '')
-    password = request.json.get('password', '')
-
-    user = Users.query.filter_by(username=username).first()
-
-    if not user:
-        return jsonify({"Message":"Wrong username!, Please try again later"}, 401)
-
-    if user.password == password:
-        token = jwt.encode({'public_id': user.id,
-                            'exp': datetime.utcnow() + timedelta(hours=8)}, Config.SECRET)
-
-        return jsonify({'token': token.decode('UTF-8'),
-                        'Message': 'Login successful!'}), 200
-    return jsonify({"Message":"Invalid password!, Please try again later"}, 401)
+    if not userlist:
+        return jsonify({"Message":"No registered user found!"}), 204
+    else:
+        return jsonify(userlist), 200
 
 
 @auth.verify_password
@@ -96,14 +86,9 @@ def verify_password(username, password):
     g.user = user
     return True
 
-@app.errorhandler(404)
-def not_found():
-    """function to format error response"""
-    return make_response(jsonify({'error': 'Not found'}), 404)
-
 @app.route('/recipe/api/v1.0/category', methods=['POST'])
 @auth.login_required
-def new_category():
+def create_new_category():
     """function to create new recipe category of a user
     ---
     tags:
@@ -125,17 +110,14 @@ def new_category():
         cat_exits = Category.query.filter_by(name=category_name).first()
 
         if cat_exits:
-            response = jsonify({"message":"Category {} already exits".format(category_name)}, 200)
+            return jsonify({"message":"Category {} already exits".format(category_name)}), 400
         else:
-
             category = Category(user_id, category_name, description)
             category.save()
 
-            response = jsonify({"message":"category {} was added successfully!".format(category.name)}, 201)
+            return jsonify({"message":"category {} was added successfully!".format(category.name)}), 201
     else:
-        response = jsonify({"message":"Please enter all details!"}, 200)
-
-    return response
+        return jsonify({"message":"Please enter all details!"}), 400
 
 
 @app.route('/recipe/api/v1.0/category/<int:category_id>', methods=['PUT'])
@@ -165,7 +147,7 @@ def update_category(category_id):
     category = Category.query.filter_by(id=category_id).first()
 
     if not category:
-        return jsonify({"Message":"No category with id {} was found!".format(category_id)}, 400)
+        return jsonify({"Message":"No category with id {} was found!".format(category_id)}), 400
 
     if category_name and description:
 
@@ -173,13 +155,10 @@ def update_category(category_id):
         category.description = description
         category.date_modified = datetime.utcnow()
         category.save()
-
-        response = jsonify({"Message": "category {} was updated successfully".format(category.id)}, 201)
+        return jsonify({"Message": "category {} was updated successfully".format(category.id)}), 201
     else:
-        response = jsonify({"Message": "Please enter new details!"}, 200)
+        return jsonify({"Message": "Please enter new details!"}), 400
 
-
-    return response
 
 @app.route('/recipe/api/v1.0/category/', methods=['GET'])
 @auth.login_required
@@ -210,15 +189,23 @@ def view_category():
             resource_for_url='view_category',
             key_name='results',
             schema=category_schema)
+
         result = pagination_helper.paginate_query()
-        return jsonify({'categories':result})
+        if result['count']:
+            return jsonify({'categories':result}), 200
+        else:
+            return jsonify({'Message':"No record matches search term"}), 204
+
     result = pagination_helper.paginate_query()
-    return jsonify({'categories':result})
+    if result['count']:
+        return jsonify({'categories':result}), 200
+
+    return jsonify({"Message":"No record found"}), 404
 
 
 @app.route('/recipe/api/v1.0/category/<int:category_id>', methods=['GET'])
 @auth.login_required
-def view_category_by(category_id):
+def view_category_by_id(category_id):
     """function to query a recipe category of a user by category id
     ---
     tags:
@@ -232,30 +219,29 @@ def view_category_by(category_id):
       200:
         description: Ok
     """
-    user = Users.query.filter_by(id=g.user.id).first()
-    categorylists = Category.query.filter_by(id=category_id).first()
+    if category_id > 0:
+        user = Users.query.filter_by(id=g.user.id).first()
+        categorylists = Category.query.filter(Category.id == category_id, Category.user_id == user.id).first()
 
-    if not categorylists:
-        response = jsonify({"Message": "No category with id {} was found!".format(category_id)}, 204)# no content found
-        return response
-
-    results = {}
-
-    if user.id == categorylists.user_id:
-
-        if categorylists:
-            results["id"] = categorylists.id
-            results["name"] = categorylists.name
-            results["user"] = categorylists.user_id
-            results["description"] = categorylists.description
-            results["date_modified"] = categorylists.date_modified
-
-            response = jsonify(results)
+        if not categorylists:
+            response = jsonify({"Message": "No category with id {} was found!".format(category_id)}), 204# no content found
             return response
-    else:
-        response = jsonify({"Message": "You don't have the right to view that category"}, 401)#unauthorized code
-        return response
-    return jsonify({"Message":"No record found!"}, 400)#bad request code
+
+        results = {}
+
+        if user.id == categorylists.user_id:
+
+            if categorylists:
+                results["id"] = categorylists.id
+                results["name"] = categorylists.name
+                results["user"] = categorylists.user_id
+                results["description"] = categorylists.description
+                results["date_modified"] = categorylists.date_modified
+
+                response = jsonify(results), 200
+                return response
+
+    return jsonify({"Message":"No record found!"}), 404# not found code
 
 
 @app.route('/recipe/api/v1.0/category/<int:category_id>', methods=['DELETE'])
@@ -278,14 +264,13 @@ def delete_category(category_id):
         category = Category.query.filter_by(id=category_id).first()
 
         if not category:
-            return jsonify({"Message":"No category with id {} was found!".format(category_id)}, 204)# no content found
+            return jsonify({"Message":"No category with id {} was found!".format(category_id)}), 204# no content found
         else:
             category.delete()
-            response = jsonify({"Message": "category {} was deleted successfully".format(category.name)}, 200)#ok
+            response = jsonify({"Message": "category {} was deleted successfully".format(category.name)}), 200#ok
             return response
     else:
-        return jsonify({"Message":"Invalid category id {}".format(category_id)}, 400)#bad request
-    return jsonify({"Message":"No record found!"}, 400)#bad request
+        return jsonify({"Message":"Invalid category id {}".format(category_id)}), 404#bad request
 
 
 @app.route('/recipe/api/v1.0/category/recipes', methods=['POST'])
@@ -314,16 +299,13 @@ def new_recipe():
     if recipe_name and ingredients and category_id > 0:
         recipe = Recipe.query.filter_by(name=recipe_name).first()
         if not recipe:
-
             recipe = Recipe(category_id, recipe_name, ingredients)
             recipe.save()
-            response = jsonify({"message": "recipe {} was added successfully!".format(recipe.name)}, 201)
-
+            return jsonify({"message": "recipe {} was added successfully!".format(recipe.name)}), 201
         else:
-            response = jsonify({"message":"Recipe name already exits"}, 200)
+            return jsonify({"message":"Recipe name already exits"}), 400
     else:
-        response = jsonify({"message":"Please enter all details!"}, 200)
-    return response
+        return jsonify({"message":"Invalid values"}), 400
 
 
 @app.route('/recipe/api/v1.0/category/recipes/<int:recipe_id>', methods=['PUT'])
@@ -353,16 +335,18 @@ def update_recipe(recipe_id):
 
     recipe = Recipe.query.filter_by(id=recipe_id).first()
 
-    if recipe:
+    if not recipe:
+        return jsonify({"Message":"No recipe with id {} was found!".format(recipe_id)}), 400
+
+    if recipe_name and ingredients:
         recipe.name = recipe_name
         recipe.ingredients = ingredients
         recipe.date_modified = datetime.utcnow()
         recipe.save()
-        response = jsonify({"message": "recipe {} was updated successfully!".format(recipe.id)}, 201)
+        return jsonify({"message": "recipe {} was updated successfully!".format(recipe.id)}), 201
     else:
-        response = jsonify({"message": "No recipes with id {} was found!".format(recipe_id)}, 401)#
+        return jsonify({"message": "No recipes with id {} was found!".format(recipe_id)}), 204#
 
-    return response
 
 
 @app.route('/recipe/api/v1.0/category/recipes/', methods=['GET'])
@@ -393,10 +377,16 @@ def view_recipe():
             key_name='results',
             schema=recipe_schema)
         result = pagination_helper.paginate_query()
-        return jsonify({'recipes':result})
-
-    result = pagination_helper.paginate_query()
-    return jsonify({'recipes':result}, 200)
+        if result['count']:
+            return jsonify({'recipes':result}), 200
+        else:
+            return jsonify({"Message": 'No record matches search term'}), 204
+    else:
+        result = pagination_helper.paginate_query()
+        if result['count']:
+            return jsonify({'recipes':result}), 200
+        else:
+            return jsonify({"Message":'No record found'}), 404
 
 
 @app.route('/recipe/api/v1.0/category/recipes/<int:category_id>', methods=['GET'])
@@ -420,44 +410,26 @@ def view_recipe_by_category(category_id):
         description: Unauthorized user
     """
     user = Users.query.filter_by(id=g.user.id).first()
-    categorylists = Category.query.filter(Category.user_id == user.id, Category.id == category_id).first()
-    recipes = Recipe.query.filter_by(category_id = category_id)
 
-    if categorylists is None:
-        response = jsonify({"Message": "No category with id {} was found!".format(category_id)}, 204)#no content found
-        return response
+    if category_id > 0:
+        categorylists = Category.query.filter(Category.user_id == user.id, Category.id == category_id).first()
+        recipes = Recipe.query.filter_by(category_id=category_id)
 
-    if recipes is None:
+        if not categorylists or not recipes:
+            return jsonify({"Message": "No category with id {} was found!".format(category_id)}), 204#no content found
 
-        response = jsonify({"Message": "No recipes for category with id {} was found!".format(category_id)}, 204)
-        return response
-
-    results = {}
-
-    if user.id == categorylists.user_id:
-
-        if recipes:
-            for recipe in recipes:
-                obj = {
-                    'id': recipe.id,
-                    'name': recipe.name,
-                    'category': recipe.category_id,
-                    'ingredients': recipe.ingredients,
-                    'date_modified':recipe.date_modified
-                    }
+        results = {}
+        for recipe in recipes:
+            if recipe:
                 results["id"] = recipe.id
                 results["name"] = recipe.name
                 results["category"] = recipe.category_id
                 results["ingredients"] = recipe.ingredients
                 results["date_modified"] = recipe.date_modified
 
-            response = jsonify(results, 200)
+                return jsonify(results), 200
 
-            return response
-    else:
-        response = jsonify({"Message": "You don't have the right to view that recipe"}, 401)#unauthorized user
-        return response
-    return jsonify({"Message":"No records found!"}, 204)
+    return jsonify({"Message":"Invalid category id!"}), 400
 
 
 @app.route('/recipe/api/v1.0/category/<int:category_id>/recipe/<int:recipe_id>', methods=['GET'])
@@ -486,43 +458,23 @@ def view_recipe_by_id(category_id, recipe_id):
     """
 
     user = Users.query.filter_by(id=g.user.id).first()
-    categorylists = Category.query.filter(Category.user_id == user.id, Category.id == category_id).first()
-    recipes = Recipe.query.filter(Recipe.id == recipe_id, Recipe.category_id == categorylists.id)
 
+    if category_id > 0:
+        categorylists = Category.query.filter(Category.user_id == user.id, Category.id == category_id).first()
+        recipes = Recipe.query.filter(Recipe.id == categorylists.id, Recipe.id == recipe_id)
 
-    if not categorylists:
-        response = jsonify({"Message": "No category with id {} was found!".format(category_id)}, 204)#no content found
-        return response
+        results = {}
+        for recipe in recipes:
+            if recipe:
+                results["id"] = recipe.id
+                results["name"] = recipe.name
+                results["category"] = recipe.category_id
+                results["ingredients"] = recipe.ingredients
+                results["date_modified"] = recipe.date_modified
 
-    if not recipes:
-
-        response = jsonify({"Message": "No recipe with id {} was found!".format(recipe_id)}, 204)
-        return response
-
-    results = []
-
-    if user.id == categorylists.user_id:
-
-        if recipes:
-            for recipe in recipes:
-                obj = {
-                    'id': recipe.id,
-                    'name': recipe.name,
-                    'category': recipe.category_id,
-                    'ingredients': recipe.ingredients,
-                    'date_modified':recipe.date_modified
-                    }
-                results.append(obj)
-
-            response = jsonify(results, 200)
-
-
-            return response
-    else:
-        response = jsonify({"Message": "You don't have the right to view that recipe"}, 401)#unauthorized user
-        return response
-    return jsonify({"Message":"No records found!"}, 204)
-
+                return jsonify(results), 200
+    
+    return jsonify({"Message":"Invalid category id!"}), 400
 
 @app.route('/recipe/api/v1.0/category/recipes/<int:recipe_id>', methods=['DELETE'])
 @auth.login_required
@@ -542,13 +494,12 @@ def delete_recipe(recipe_id):
       204:
         description: No content found
     """
-    recipe = Recipe.query.filter_by(id=recipe_id).first()
-    if not recipe:
-        return jsonify({"Message":"No recipe with id {} was found!".format(recipe_id)}, 204)#no content found
-    else:
-        recipe.delete()
-        response = jsonify({"Message": "recipe {} deleted successfully!".format(recipe_id)}, 200)
-    return response
+    if recipe_id > 0:
+        recipe = Recipe.query.filter_by(id=recipe_id).first()
+        if not recipe:
+            return jsonify({"Message":"No recipe with id {} was found!".format(recipe_id)}), 204#no content found
+        else:
+            recipe.delete()
+            return jsonify({"Message": "recipe {} deleted successfully!".format(recipe_id)}), 200
 
-if __name__ == '__main__':
-    app.run()
+    return jsonify({"Message":"Invalid id"}), 400
